1.  Keeping the socket open on my system, when apcupsd is shutdown, collectd dies as well. I re-inserted the net_close() and now when apcupsd is shutdown (or otherwise goes away) collectd stays running.
2. Changed the 'complain' and 'complain_step' code so that it will chirp the first time, then
every 6 hours. As it was originally it would only start complaining after 6 hours. Additionally
I made complain an unsigned int. If it was an int and rolled over I wasn't certain what would happen.

3. 'global_host' set to localhost. It was set to NULL, but what happens is if someone does not enter
an apcups section to collectd.conf, then it will not connect to anyone. With 'global_host' set to localhost
the default behavior is now to connect to the apcupsd on the localhost.

4. Added entry for apcupsd.so in 'contrib/collectd.spec'.