More improvements on the named/dns plugin.
The `dnstop.c' code now contains my IPv6 patch to capture IPv6 traffic. The
plugin captures both, queries and replies, and collects statistics
accordingly. This means that the qtype of queries and the rcode of replies are
collected. The plugin can now be configured to ignore certain sources, but
this functionality is to be improved.