ping plugin: Add the hosts during initialization, not when reading the config file.
Since adding a host to liboping might block when no nameserver can be reached
or another problem occures (normally this means network misconfiguration) it's
not the best idea to do this when reading the config file. At that point we
didn't fork to the background yet, so the boot process may hang which is not good.

Also, hosts that cannot be added right away are tried again in increasing intervals.