iptables plugin: Adds a plugin to collect iptables'-counters.
My first attempt at a module for this is an iptables modules This will try to
read out the rules from iptables for a specific table and chain.
It will go through the rules looking for ones with "comment" matches and will
use the comment as a base for the generated filename.

So basically looking for lines like
 $ iptables -t mangle -A incoming -p tcp -m comment --comment "tcp"
will generate a iptables-incoming/tcp.rrd file from the byte counters

There used to also be an account iptables module but that seems to have been
neglected, which would generate /proc entries for specific rules, don't know
what would work better. Either way, this seems to work somewhat.