configure.in: Generate version dynamically using version-gen.sh.
This dynamically creates the version string when the configure script is
created. If run from within a repository the output of "git describe" is
used to create the version string. Else it defaults to a hardcoded value
which should be something like "$latest_version.git".

The idea is to update the default version to the new version (without
".git" appended) as the last commit of a release cycle. This commit is
tagged as the new version. The next commit after the release will set the
default version to "$new_version.git".

The benefit is that a pre-release with a quite well defined and unique
version can be easily created at any time.

Signed-off-by: Sebastian Harl <sh@tokkee.org>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>