Fixed Fedora specfile for Fedora 7 and collectd 4.0.5
Hi,

Here is a working spec file for Fedora 7.
Copy the tar.gz file to /usr/src/redhat/SOURCES and run
  rpmbuild -ba collectd.spec
It has only been tested on  x86_64, but I well see if I can find time to test
on x86 as well.

Upgrade not tested.

The file /etc/collection.conf referenced in collection.cgi is missing in the
tar file and not installed by the spec file.

Regards
Kjell Randa