contrib/fedora/collectd.spec: Updated for 4.0.6.
Hi,

Here is an updated version of the F7 rpm spec file.
It now handles updates as well (rpm -U).

When new versions of collectd are out please bump the version tag at the top of
the spec file.
It will then build correct version of collectd.

Regards
Kjell Randa

Signed-off-by: Florian Forster <octo@huhu.verplant.org>