contrib/collectd2html.pl: Allow generation of graphs from remote hosts or in an alternative directory.
In the contribs area of the package there is a nice collectd2html.pl script
which assumes that the data is present in /var/lib/collectd and infers the
hostname. All fine and dandy, but there is no way to override this. This is
useful if one generates the data on a machine, but generates the graphs on
another one.

With the attached patch now is possible to specify a different host and/or a
different directory for the data.

The added options are '--host-is' and '--data-dir'.

An invocation for a remote host whose rrd files have been placed in
/var/tmp/collectd-remoteh looks like:
 collectd2html.pl --host-is=remoteh --data-dir=/var/tmp/collectd-remoteh