netlink plugin: Added a plugin for (very) enhanced Linux traffic statistics.
This plugin uses the netlink sockets provided by Linux 2.2 and later to get
information about the links, qdiscs, classes, and filters in the network stack
of a Linux system. With this plugin it's possible to get some _very_ detailed
statistics.

Still to do: Update the `types.db', `collectd.conf(5)', and `configure.conf.in'.