netlink plugin: Fix several issues, including the collection of classes and filters.
When requesting qdiscs for ifindex == 0 you get the qdiscs of all interfaces.
The same doesn't work for classes and filters.
This commit changes the behavior to iterate over a list of interfaces and
request qdiscs, classes and filters for each interface seperately. The
ignorelist is checked before sending a request to avoid unnecessary requests.

Newlines have been removed from error messages.