rrdtool plugin: Use `-b' (begin) when creating rrd-files.
When receiving data from a host that has a clock that lacks behind the servers
clock, the newly created rrd-files had a timestamp that was ten seconds in the
past from the server's point of view. If the client was, for example, 10:00
minutes late, there would therefore be 9:50 minutes of invalid data which cause
warnings, empty graphs and overall missery. This commit changes this behavior
to not use the default `begin' time but set it explicitely based on the
client's time.