perl plugin: Improved value list conversion.
The perl plugin used to store an internal list of data-source types for
each data-set registered by Perl plugins to get fast access to this
information (which is required to correctly convert a Perl data-set to a
data_set_t struct). However, this prevented Perl plugins to use data-sets
defined in types.db. This patch fixes this issue by requesting the
data-set definition from collectd.

Signed-off-by: Sebastian Harl <sh@tokkee.org>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>