apache plugin: Improved some error messages and fixed-buffer string copying.
Return an error when initialization fails. This includes the case when no URL
is configured. The user is informed about this as clear as possible.

Also, `strncpy' and `snprintf' now are followed by a forced NULL-termination of
the strings to be on the save side. This also changes the plugin to look more
like the other plugins which mostly do this already.