rrdtool plugin: Reduce the filesize of files created per default.
When creating an RRD file, the rrdtool plugin doesn't do what it's supposed to
do. Instead of creating 5 RRAs it creates 4 and in doing so it skips the first
one. The result is about 2x the number of datapoints total. The fix below works
for me. It's for 4.2.2 but it looks like this hasn't changed in the latest 4.3
beta.

Resolves: #5