Add libiptc.
libiptc is not meant to be a public interface by upstream. Debian has thus
removed it from all packages. To be able to use it, we need to take care
of it ourselves.

The configure script now checks, if libiptc is available and uses the
shipped version as a fallback.

The following changes to the libiptc sources were required:
 * added copyright headers mentioning the "Netfilter Core Team" as author
 * changed libiptc/* includes to * only
 * made libiptc.c:standard_target_map() static

Signed-off-by: Sebastian Harl <sh@tokkee.org>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>