disk plugin: Use correct name under Linux 2.4
Hello everybody.

I have noticed that disk graphs are displayed by collection.cgi as
disk (71687000) instead of the correct disk name on 2.4.x linux
kernels.
The problem seems caused by the fact that the position of disk name of
/proc/partitions (used in 2.4 kernels) differs from /proc/diskstats
(used in 2.6 kernels).

I include as attachment a patch that seems to fix this problem.
I have tested it on 2.4.20 (REDHAT 9) and 2.6.24.4 (Debian Sid) systems.

Thanks a lot for Collectd. It is a _very_ useful program.

Alessandro

Signed-off-by: Florian Forster <octo@huhu.verplant.org>