qmail plugin: Added a plugin to count messages in qmail queues.
Hello.
Here is my first collectd plugin.
It shows the two values of qmail queues, messages and todo.
The latter should be zero in normal cases. I included it because it
may help diagnose problems.
Maybe a config parameter could be added for it.

It supports a parameter for the Qmail base directory:

<Plugin qmail_queue>
       QmailDir "/var/qmail"
</Plugin>

Feel free to modify it if you need.
Alessandro